Added DrRect, DrRectF. Updated license text.

+++ library/types/dr_vec2.cpp
@@ -17,11 +17,11 @@
 //####################################################################################
 //##    Constructors
 //####################################################################################
-DrVec2::DrVec2()                        { x = static_cast<float>(0);        y = static_cast<float>(0); }
-DrVec2::DrVec2(float f)                 { x = f;                            y = f; }
-DrVec2::DrVec2(float x_, float y_)      { x = x_;                           y = y_; }
-DrVec2::DrVec2(const DrVec2 &v)         { x = static_cast<float>(v.x);      y = static_cast<float>(v.y); }
-DrVec2::DrVec2(const glm::vec2 &v)      { x = static_cast<float>(v.x);      y = static_cast<float>(v.y); }
+DrVec2::DrVec2()                        { x = 0.f;                      y = 0.f; }
+DrVec2::DrVec2(float f)                 { x = f;                        y = f; }
+DrVec2::DrVec2(float x_, float y_)      { x = x_;                       y = y_; }
+DrVec2::DrVec2(const DrVec2 &v)         { x = static_cast<float>(v.x);  y = static_cast<float>(v.y); }
+DrVec2::DrVec2(const glm::vec2 &v)      { x = static_cast<float>(v.x);  y = static_cast<float>(v.y); }
 
 
 //####################################################################################
